Move messageList into its own component folder
