+++ Source/SelfService/Web/integrations/connection/routes.tsx
@@ -8,7 +8,7 @@ import { RouteObject, Navigate } from 'react-router-dom';
 import { ConnectionDetails } from './connectionDetails';
 import { ConfigurationView } from './connectionDetails/configuration';
 import { ExposeDataView } from './connectionDetails/exposeData';
-import { MessagesListView } from './connectionDetails/messages/MessagesListView';
+import { MessagesListView } from './connectionDetails/messages/messagesList';
 import { MessagesViewRoot } from './connectionDetails/messages';
 import { ChangeMessageView } from './connectionDetails/messages/changeMessage';
